split out tests properly
